requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# backwards compatibility
pathlib2

# Own packages
tensorflow == 2.6.0
keras == 2.6.0
pandas == 1.3.4
numpy == 1.21.5
lightgbm == 3.3.1
scikit-learn == 1.0.1
matplotlib
wandb

# pip install -r requirements.txt